requirements.txt
asgiref==3.10.0
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
djangorestframework==3.16.1
fonttools==4.60.1
iniconfig==2.1.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
gunicorn
whitenoise